Swears list now persists!
I've finally worked out how to get the swear list to persist between restarts of shizbot. Had to make a dummy class to hold the swear list, but it now auto-persists the changes you make to the list.

Next up: get the regex rebuilt whenever there's a change to the list, so new (or old) swears take effect (or don't)