Update set credits to store a numerical value
The demo-man command `morality set credit <user> <credits>` was blindly taking the supplied match (which is a string) and assigning it to the users `morality_credits` value. When this was then 'summed' as part of the `morality stats` command, the first string value encountered turned the total value into a string also, thus concatenating all subsequent values. This lead to wildly inaccurate numbers being generated.

I've changed the `set credit` functionality to now convert the supplied value into something numeric, so we should just need to re-run the `set` for users I've identified with string values for their `morality_credits`